+++ app/components/house-ninja/referral-code/component.js
@@ -18,8 +18,8 @@ export default class ServiceAreaComponent extends Component {
     try {
       await Share.share({
         title: 'Try House Ninja for two months FREE!', //iOS - Displays in the email subject
-        text: `Hi! I wanted to share a service with you that I think you’ll love: House Ninja. If you join today and use my code ${this.promoCode} you’ll get your first two months free!`, // iOS - Main content of email and text
-        url: 'www.houseninja.co/sign-up', // iOS appended to main content (text property)
+        text: `Hi! I wanted to share a service with you that I think you’ll love: House Ninja. If you join today and use my code ${this.promoCode} you’ll get your first two months free! www.houseninja.co/sign-up`, // iOS - Main content of email and text
+        // url: '', // iOS appended to main content (text property), but doesn't work
         // dialogTitle: 'Share with buddies', //Android only. Not sure where it presents.
       });
     } catch (e) {