+++ app/services/onboarding.js
@@ -15,6 +15,8 @@ const CACHED_MODELS = [
   'user',
 ];
 
+const TTL_MINUTES = 30;
+
 export default class OnboardingService extends Service {
   @service storage;
   @service store;
@@ -28,7 +30,7 @@ export default class OnboardingService extends Service {
 
   async completeStep(step) {
     this.currentStep = step;
-    await this.storage.setLocal('current-step', this.nextStep);
+    await this.storage.setLocal('current-step', this.nextStep, TTL_MINUTES);
     const user = this.localModel('user');
     if (isPresent(user)) {
       user.onboardingStep = this.nextStep;
@@ -48,7 +50,7 @@ export default class OnboardingService extends Service {
   }
 
   routeFromStep(step) {
-    return `signup.${step}`;
+    return `onboarding.${step}`;
   }
 
   async userFromOnboardingCode(onboardingCode) {
@@ -99,7 +101,7 @@ export default class OnboardingService extends Service {
       }
     }).compact();
     records.forEach(async (record) => {
-      await this.storage.setLocal(record.data.type, record);
+      await this.storage.setLocal(record.data.type, record, TTL_MINUTES);
     });
   }
 
@@ -108,7 +110,7 @@ export default class OnboardingService extends Service {
   }
 
   set zipcode(zip) {
-    this.storage.setLocal('zipcode', zip);
+    this.storage.setLocal('zipcode', zip, TTL_MINUTES);
   }
 
   async fetchLocalModel(modelType) {