+++ mirage/factories/user.js
@@ -1,5 +1,5 @@
 import { Factory, association } from 'miragejs';
-import faker from '@faker-js/faker';
+import { faker } from '@faker-js/faker';
 import hmacSHA256 from 'crypto-js/hmac-sha256';
 import ENV from 'houseninja/config/environment';
 import getPlatform from 'houseninja/utils/get-platform';
@@ -8,6 +8,7 @@ import { ALL as OnboardingSteps } from 'houseninja/data/enums/onboarding-step';
 
 export default Factory.extend({
   promoCode: association(),
+  subscription: association(),
 
   email() {
     return faker.internet.email();